OpenMapTiles based natural earth import
This process creates the 'natural_earth_vector.sqlite' file. This file is then imported into Postgresql (like the origional omt project did). This seemed to work, but I am trying to replace it with a script that takes the shape files and imports them into postgresql directly (skipping the sqlite like I used to do)